examples/segment_objects/requirements.txt
Pillow==9.1.1
numpy==1.21.5